github_agent_sequential/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-adk>=1.18.0
google-genai>=1.49.0
aiohttp>=3.10.0
colorlog==6.8.0
litellm>=1.57.13
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0